requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pypdf>=3.17.0
numpy>=1.26.0
tiktoken>=0.7.0
openai>=1.12.0
huggingface-hub>=0.20.0
sentence-transformers>=2.6.0
python-dotenv>=1.0.1
python-multipart>=0.0.21
faiss-cpu>=1.7.4
langchain-community>=0.2.0
langchain-core>=0.2.0
spacy>=3.7.4
pydantic-settings>=2.0.0
docx2txt>=0.8